// frontend/src/backend.d.ts
export interface NormalizedFuturesPosition {
    pnl: number;
    markPrice: number;
    leverage: number;
    positionSide: string;
    liquidationPrice: number;
    entryPrice: number;
    market: BinanceFuturesMarket;
    symbol: string;
    positionAmt: number;
}
export interface UserProfile {
    name: string;
    email?: string;
    preferences?: string;
}
export enum BinanceFuturesMarket {
    usdt_m = "usdt_m",
    coin_m = "coin_m"
}